+++ backend/src/main/java/com/example/backend/service/room/RoomAddService.java
@@ -3,6 +3,6 @@ package com.example.backend.service.room;
 import com.alibaba.fastjson2.JSONObject;
 
 public interface RoomAddService {
-    JSONObject roomAdd(String type, Integer price, Integer count, Float discount,
+    JSONObject roomAdd(String type, float price, Integer count, Float discount,
                        String location, String info);
 }
