+++ backend/src/main/java/com/example/backend/service/room/RoomQueryService.java
@@ -3,7 +3,7 @@ package com.example.backend.service.room;
 import com.alibaba.fastjson2.JSONObject;
 
 public interface RoomQueryService {
-    JSONObject roomQueryService(String type, String location, Integer lowerPrice, Integer highPrice);
+    JSONObject roomQueryService(String type, String location, Float lowerPrice, Float highPrice);
 
     JSONObject roomQueryOneService(Integer id);
 }
